Add NTJsonModelContainer protocol declaration to the extension declaration even though it's already declared in the (Private) category. (Xcode 6 fix)
Previously the class would come back as conforming to the protocol even though the (Private) category for NTJsonModelArray is never explicitly implemented. (It's methods actually live in the main class and the (Private) category is only used to hide the values to external users of the class.) In xcode6 /ios 8 (at least beta 5) it appears this no longer works. My suspicion is the class does not conform to the protocol unless the category is explicitly implemented. Adding the category here seems to fix the issue.